tpm: cap PCRs that shim touched at exit time
TPM PCRs act via a series of extensions of hashes - the final value of the
PCR depends on the value of every value measured into it, along with the
order of those measurements (eg, measuring a followed by b followed by c
will result in a different PCR value than measuring a followed by c followed
by b). However, the TPM has no real way of knowing *what* is performing any
given measurement. As a result, shim measuring a followed by b followed by c
is indistinguishable from shim measuring a followed by b, followed by OS
code measuring c. Since shim and the OS are acting in separate security
contexts, this is potentially a problem.

The TPM specification provides a solution for this, in the form of "capping".
When transitioning between contexts, an EV_SEPARATOR event containing a
fixed value is logged. As a result, the previous scenario will now look like
shim measuring a, then b, then c, then a separator - or, alternatively, shim
measuring a, then b, then a separator, then the OS measuring c. The different
position of the separator in this chain means that the final PCR value will
differ, making it possible to differentiate them.

It would be tempting to set the separator value to 0x5348494D (SHIM), but
unfortunately table 9 of the TCG PC client platform spec says:

"Per Sections 2.3.4 and 7.2, used to delimit actions taken during the pre-OS
and OS environments. For this situation, the digests field MUST contain the
tagged hash of the event data for each PCR bank. The event data size MUST be
4.  The event field MUST contain the hex value 00000000h or FFFFFFFFh."

So stick with zero.

We transition out of Shim in two separate scenarios - if shim exits, or if
ExitBootServices() is called. It makes sense to cap in both of these.

Signed-off-by: Matthew Garrett <mgarrett@aurora.tech>